added a check to remove the unnecessary error in powershell

+++ build-tools/downloadNuget.ps1
@@ -1,4 +1,6 @@
-Remove-Item ../source/.nuget/NuGet.exe
+if (Test-Path ../source/.nuget/NuGet.exe) {
+  Remove-Item ../source/.nuget/NuGet.exe
+}
 (new-object System.Net.WebClient).DownloadFile('http://nuget.org/nuget.exe','../source/.nuget/NuGet.exe')
 cd ../source
 ./.nuget/NuGet.exe restore "FloydPink.Flickr.Downloadr.sln" -verbosity detailed
